endiannes

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -6,7 +6,7 @@
  */
 int get_endianness(void)
 {
-	unsigned int i = 1;
+	unsigned long int i = 1;
 	
 	return (*(char *)&n);
 }
